creates setup configuration for test database in container
