+++ server/jest-setup.js
@@ -1,0 +1,5 @@
+const { start } = require("./jest-mssql-container");
+
+module.exports = async () => {
+  await start();
+};
