+++ server/jest-teardown.js
@@ -1,0 +1,5 @@
+const { stop } = require("./jest-mssql-container");
+
+module.exports = async () => {
+  await stop();
+};
